Ensure ImportVolume is idempotent
The ImportVolume method must sync the tarball
extraction and import to avoid multiple goroutines
trying to do so across several launch or refresh
tasksets which use the same SDK.
