+++ internal/workshop/lxd/lxd_backend.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"net/http"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"runtime"
 	"sync"
@@ -40,6 +39,9 @@ var (
 	// instance map with the LXD backend.
 	imageLock        sync.Mutex
 	currentDownloads map[string]*downloadOp
+
+	volumes     sync.Mutex
+	volumeLocks map[string]chan struct{}
 )
 
 //go:embed start_command.sh
@@ -51,6 +53,10 @@ func init() {
 	if currentDownloads == nil {
 		currentDownloads = make(map[string]*downloadOp)
 	}
+
+	volumes.Lock()
+	defer volumes.Unlock()
+	volumeLocks = make(map[string]chan struct{})
 }
 
 type Backend struct {
@@ -639,185 +645,6 @@ func (s *Backend) WorkshopFs(ctx context.Context, name string) (workshop.Worksho
 	}
 
 	return workshop.NewWorkshopFs(sftp), nil
-}
-
-func (s *Backend) CreateVolume(ctx context.Context, name string) error {
-	conn, err := s.LxdClient(ctx)
-	if err != nil {
-		return err
-	}
-	defer conn.Disconnect()
-
-	// Create the storage volume entry
-	vol := api.StorageVolumesPost{}
-	vol.Name = name
-	vol.Type = "custom"
-	vol.ContentType = "filesystem"
-	vol.Config = map[string]string{}
-
-	err = conn.CreateStoragePoolVolume(storagePool, vol)
-	if api.StatusErrorCheck(err, http.StatusConflict) {
-		return workshop.ErrVolumeAlreadyExists
-	}
-	return err
-}
-
-var volumeIndex = `name: %s
-backend: %s
-type: custom
-config:
-  volume:
-    name: %s
-    description: "SDK Volume"
-    type: custom
-    content_type: filesystem
-`
-
-func volumeIndexContent(name string) string {
-	return fmt.Sprintf(volumeIndex, name, storagePool, name)
-}
-
-func (s *Backend) ImportVolume(ctx context.Context, name string, tarball string) error {
-	conn, err := s.LxdClient(ctx)
-	if err != nil {
-		return err
-	}
-	defer conn.Disconnect()
-
-	_, _, err = conn.GetStoragePoolVolume(storagePool, "custom", name)
-	if err == nil {
-		return workshop.ErrVolumeAlreadyExists
-	}
-	if !api.StatusErrorCheck(err, http.StatusNotFound) {
-		return err
-	}
-
-	// The tarballs will be transformed into a LXD-compatible backup format to
-	// create them directly as a custom volume. The LXD's tar archive has the
-	// following format:
-	//
-	// backup/
-	//	volume/
-	//  index.yaml
-
-	dir, err := os.MkdirTemp("", name)
-	if err != nil {
-		return err
-	}
-	defer os.RemoveAll(dir)
-
-	unpack := exec.CommandContext(ctx, "tar",
-		"--extract",
-		"--file="+tarball,
-		"--transform",
-		"s,^,volume/,",
-		"--directory="+dir,
-	)
-
-	if _, err := unpack.Output(); err != nil {
-		logger.Debugf("Failed to unpack volume tarball: %v", err)
-		return err
-	}
-
-	// Generate index.yaml for the volume.
-	if err = os.WriteFile(filepath.Join(dir, "index.yaml"), []byte(volumeIndexContent(name)), 0644); err != nil {
-		return err
-	}
-
-	newtar := filepath.Join(dir, filepath.Base(tarball))
-
-	// Read the metadata to store it as a volume's property. This is not ideal
-	// when the backend knows the name of the file with metadata as the volume
-	// manager should be able to import any tarball as a volume. But given that
-	// it is only applicable to SDKs in the nearest future, it should be acceptable
-	// as the alternative would be to change the interface to accept the metadata.
-	meta, err := os.ReadFile(filepath.Join(dir, "volume", "meta", "sdk.yaml"))
-	if err != nil {
-		return err
-	}
-
-	repack := exec.CommandContext(ctx, "tar",
-		"--remove-files",
-		"--create",
-		"--file",
-		newtar,
-		"--transform",
-		"s,^,backup/,",
-		"--directory="+dir,
-		"--no-same-owner",
-		"volume/",
-		"index.yaml",
-	)
-
-	if _, err := repack.Output(); err != nil {
-		logger.Debugf("Failed to repack volume tarball: %v", err)
-		return err
-	}
-
-	f, err := os.Open(newtar)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	vol := lxd.StoragePoolVolumeBackupArgs{
-		BackupFile: f,
-		Name:       name,
-	}
-
-	op, err := conn.CreateStoragePoolVolumeFromBackup(storagePool, vol)
-	if err != nil {
-		return err
-	}
-
-	if err = op.WaitContext(ctx); err != nil {
-		return err
-	}
-
-	return conn.UpdateStoragePoolVolume(storagePool, "custom", name, api.StorageVolumePut{
-		Config: map[string]string{
-			workshop.ConfigVolumeMeta: string(meta),
-		}}, "")
-}
-
-func (s *Backend) AttachVolume(ctx context.Context, wp, name, what string) error {
-	return s.AddWorkshopMount(ctx, wp, workshop.Mount{Name: name, What: what, Where: name, Type: workshop.Volume})
-}
-
-func (s *Backend) DetachVolume(ctx context.Context, wp, name string) error {
-	return s.RemoveWorkshopMount(ctx, wp, name)
-}
-
-func (s *Backend) DeleteVolume(ctx context.Context, name string) error {
-	conn, err := s.LxdClient(ctx)
-	if err != nil {
-		return err
-	}
-	defer conn.Disconnect()
-
-	return conn.DeleteStoragePoolVolume(storagePool, "custom", name)
-}
-
-func (s *Backend) Volume(ctx context.Context, name string) (workshop.VolumeInfo, error) {
-	var info workshop.VolumeInfo
-	conn, err := s.LxdClient(ctx)
-	if err != nil {
-		return info, err
-	}
-	defer conn.Disconnect()
-
-	vol, _, err := conn.GetStoragePoolVolume(storagePool, "custom", name)
-	if api.StatusErrorCheck(err, http.StatusNotFound) {
-		return info, workshop.ErrVolumeNotFound
-	}
-
-	if err != nil {
-		return info, err
-	}
-
-	info.Name = vol.Name
-	info.Config = vol.Config
-	return info, nil
 }
 
 func (s *Backend) LxdClient(ctx context.Context) (lxd.InstanceServer, error) {
