+++ tests/main/launch-multiple/.workshop/ws-one.yaml
@@ -1,0 +1,5 @@
+name: ws-one
+base: ubuntu@22.04
+sdks:
+  - name: test-sdk-basic
+    channel: latest/stable
