+++ tests/main/launch-multiple/.workshop/ws-three.yaml
@@ -1,0 +1,5 @@
+name: ws-three
+base: ubuntu@22.04
+sdks:
+  - name: test-sdk-basic
+    channel: latest/stable
