+++ tests/main/launch-multiple/.workshop/ws-two.yaml
@@ -1,0 +1,5 @@
+name: ws-two
+base: ubuntu@22.04
+sdks:
+  - name: test-sdk-basic
+    channel: latest/stable
